+++ lib/data/datasource/custom_cups_local_datasource.dart
@@ -1,0 +1,6 @@
+import 'package:hidroly/model/water_button.dart';
+
+abstract class CustomCupsLocalDataSource {
+  Future<void> createCustomCup(WaterButton waterButton);
+  Future<List<WaterButton>> getAllCustomCups();
+}
