+++ lib/data/datasource/user_local_datasource.dart
@@ -1,0 +1,7 @@
+import 'package:hidroly/model/user.dart';
+
+abstract class UserLocalDataSource {
+  Future<void> createUser(User user);
+  Future<void> updateUser(User user);
+  Future<User?> getUser(int userId);
+}
